+++ NEWS.md
@@ -1,5 +1,6 @@
 # variancePartition 1.27.14
  - fix bug in `topTable()`
+ - update docs
 
 # variancePartition 1.27.13
  - add `diffVar()` test of differential variance